fuzz: move `AddrManDeterministic` to util
